Scan fix: cs #35, implicit narrowing

+++ engine-api/src/main/java/io/nosqlbench/engine/api/activityapi/cyclelog/outputs/ReorderingConcurrentResultBuffer.java
@@ -40,8 +40,8 @@ public class ReorderingConcurrentResultBuffer implements Output, CanFilterResult
     private final LinkedList<CycleResultsSegment> segments = new LinkedList<>();
     private final Output downstream;
     private final int threshold;
-    private int currentCount;
-    private int segmentCount;
+    private long currentCount;
+    private long segmentCount;
     private Predicate<ResultReadable> resultFilter;
 
     public ReorderingConcurrentResultBuffer(Output downstream) {
